requirements.txt
gradio>=6.0.2
matplotlib>=3.10.7
numpy>=2.3.5
pandas>=2.3.3
Pillow>=12.0.0
protobuf>=6.33.1
python-dotenv>=1.2.1
torch>=2.9.1
torch>=2.6.0+cu124
torch>=2.9.0
torchvision>=0.24.1
torchvision>=0.24.0
